chore(codegen): regenerate scalar SQL + bindings under SQL-standard names
Regenerated src/v3/scalars, eql-bindings Rust/TS/JSON from the renamed
catalog; removed stale int4/int2/int8/bool/float4/float8 surface. Updated
hand-written mod.rs module decls, v3_conformance/catalog_parity tests,
domain_type/lib doc examples, README, and SEM/schema.sql comments.

+++ src/v3/schema.sql
@@ -2,7 +2,7 @@
 --! @brief EQL v3 schema creation
 --!
 --! Creates the eql_v3 and eql_v3_internal schemas. eql_v3 is the public API:
---! the self-contained encrypted-domain type families (eql_v3.int4, eql_v3.int8,
+--! the self-contained encrypted-domain type families (eql_v3.integer, eql_v3.bigint,
 --! and future scalar domains) — their jsonb-backed domains, index-term
 --! extractors, aggregates, AND the operator-backing comparison wrappers
 --! (eq/neq/lt/lte/gt/gte/contains/contained_by, plus the jsonb containment
